Cosine distance between vector differences
Instead of cos. dist. between the search and a candidate query,
it computes cos. dist. between two difference vectors:

 - 1st diff vector = <matched words> - <missing words>
 - 2nd diff vector = <matched words> - <non-matched words>

where: matched words are words that can be found in both the search
query and the candidate query, missing words are words present only
in the search query and non-matched words are words present only in
the candidate query.

The difference between matched and non-matched words vectors in a similar
query (a vector itself) should be close to the difference between matched
and missing words vectors in the search query.